backend/requirements.txt
# 北京看房地图 - Python 依赖

# Web框架
fastapi==0.115.6
uvicorn[standard]==0.34.0

# 数据库
pymysql==1.1.1
sqlalchemy==2.0.36

# 其他工具
python-multipart==0.0.18
python-dotenv==1.0.1
